--- src/features/auth/api/index.ts
@@ -1,3 +1,0 @@
-export * from './postLogin';
-export * from './postSignup';
-export * from './postRefreshToken';
