--- src/features/auth/model/index.ts
@@ -1,4 +1,0 @@
-export * from './useLogin';
-export * from './useSignUp';
-export * from './useAuthStore';
-export * from './types';
