+++ src/features/auth/model/useAuthStore.ts
@@ -1,5 +1,6 @@
 import { create } from 'zustand';
 import { persist } from 'zustand/middleware';
+
 import type { AuthState } from './types';
 
 const initialState = {
