+++ src/features/auth/ui/GoogleLoginButton.tsx
@@ -1,6 +1,7 @@
-import { Button } from '@/shared/ui';
-import googleIcon from '@/assets/google.svg';
 import { useNavigate } from 'react-router';
+
+import googleIcon from '@/assets/google.svg';
+import { Button } from '@/shared/ui';
 
 export const GoogleLoginButton = () => {
   const navigate = useNavigate();
